BW-1: Implemented main application logic

+++ store/src/main/java/com/brandwatch/store/domain/service/impl/ProductServiceImpl.java
@@ -6,6 +6,8 @@ import com.brandwatch.store.domain.service.ProductService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -15,21 +17,42 @@ public class ProductServiceImpl implements ProductService {
     private final ProductRepository productRepository;
 
     @Override
+    public List<Product> findAll() {
+        return productRepository.findAll();
+    }
+
+    @Override
     public Optional<Product> findById(UUID id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Id must be not null when finding by id");
+        }
         return productRepository.findById(id);
     }
 
     @Override
+    public Map<UUID, Product> findAllAsMap() {
+        return productRepository.findAllAsMap();
+    }
+
+    @Override
     public Product create(Product product) {
-        if (product.getId() != null) {
+        if (product == null || product.id() != null) {
             throw new IllegalArgumentException("Id must be null when creating product");
         }
         return productRepository.save(product);
     }
 
     @Override
+    public List<Product> updateAll(Iterable<Product> products) {
+        if (products == null) {
+            throw new IllegalArgumentException("Products list cannot be null when updating all");
+        }
+        return productRepository.saveAll(products);
+    }
+
+    @Override
     public Product update(Product product) {
-        if (product.getId() == null) {
+        if (product == null || product.id() == null) {
             throw new IllegalArgumentException("Id must be not null when updating product");
         }
         return productRepository.save(product);
@@ -37,6 +60,8 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteById(UUID id) {
-        productRepository.deleteById(id);
+        if (id != null) {
+            productRepository.deleteById(id);
+        }
     }
 }
